resource-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
flask-cors==4.0.0
